+++ pairDna.h
@@ -13,7 +13,7 @@ public:
     ~pairDna(){};
 
     unsigned long getLength()const;
-    Nucleotide operator[](int)const;
+    Nucleotide operator[](size_t)const;
 
     int getIdSeq()const;
     std::string getNameSeq()const;
